Some Updates

+++ src/SimpleStorage.sol
@@ -2,29 +2,36 @@
 pragma solidity ^0.8.19;
 
 contract SimpleStorage {
-    uint FavoriteNumber;
+    uint256 FavoriteNumber;
+
     struct Person {
-        uint favoriteNumber;
+        uint256 favoriteNumber;
         string name;
     }
+
     Person[] public listOfPerson;
-    function  getListOfPerson(uint index)public view returns (Person memory) {
+
+    function getListOfPerson(uint256 index) public view returns (Person memory) {
         return listOfPerson[index];
     }
-    mapping(string => uint) public nameToFavoriteNumber;
+
+    mapping(string => uint256) public nameToFavoriteNumber;
+
     constructor() {
-        FavoriteNumber=0;
+        FavoriteNumber = 0;
     }
-    function store(uint _FavoriteNumber) public {
-        FavoriteNumber=_FavoriteNumber;
+
+    function store(uint256 _FavoriteNumber) public {
+        FavoriteNumber = _FavoriteNumber;
     }
-    function retrieve() public view returns (uint){
+
+    function retrieve() public view returns (uint256) {
         return FavoriteNumber;
     }
-    function addPerson(string memory _name,uint _favNumber)  public {
-        listOfPerson.push(Person(_favNumber,_name));
-        nameToFavoriteNumber[_name]=_favNumber;
+
+    function addPerson(string memory _name, uint256 _favNumber) public {
+        listOfPerson.push(Person(_favNumber, _name));
+        nameToFavoriteNumber[_name] = _favNumber;
     }
 }
-// forge script script/DeploySimpleStorage.s.sol --rpc-url $RPC_URL --broadcast --private-key $PRIVATE_KEY
-// 0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9
+// SEP - 0x52d273ACB6a992276a9eA8077Bd3193f076e399C
